Add new column type.
